add var_pop
