+++ core/src/main/java/io/questdb/griffin/engine/functions/groupby/VarPopGroupByFunctionFactory.java
@@ -25,11 +25,14 @@
 package io.questdb.griffin.engine.functions.groupby;
 
 import io.questdb.cairo.CairoConfiguration;
+import io.questdb.cairo.map.MapValue;
 import io.questdb.cairo.sql.Function;
+import io.questdb.cairo.sql.Record;
 import io.questdb.griffin.FunctionFactory;
 import io.questdb.griffin.SqlExecutionContext;
 import io.questdb.std.IntList;
 import io.questdb.std.ObjList;
+import org.jetbrains.annotations.NotNull;
 
 public class VarPopGroupByFunctionFactory implements FunctionFactory {
     @Override
@@ -46,4 +49,60 @@ public class VarPopGroupByFunctionFactory implements FunctionFactory {
     public Function newInstance(int position, ObjList<Function> args, IntList argPositions, CairoConfiguration configuration, SqlExecutionContext sqlExecutionContext) {
         return new VarPopGroupByFunction(args.getQuick(0));
     }
+
+    private static class VarPopGroupByFunction extends AbstractStdDevGroupByFunction {
+
+        public VarPopGroupByFunction(@NotNull Function arg) {
+            super(arg);
+        }
+
+        @Override
+        public double getDouble(Record rec) {
+            long count = rec.getLong(valueIndex + 2);
+            if (count > 0) {
+                double sum = rec.getDouble(valueIndex + 1);
+                return sum / count;
+            }
+            return Double.NaN;
+        }
+
+        @Override
+        public String getName() {
+            return "var_pop";
+        }
+
+
+        // Chan et al. [CGL82; CGL83]
+        @Override
+        public void merge(MapValue destValue, MapValue srcValue) {
+            double srcMean = srcValue.getDouble(valueIndex);
+            double srcSum = srcValue.getDouble(valueIndex + 1);
+            long srcCount = srcValue.getLong(valueIndex + 2);
+
+            double destMean = destValue.getDouble(valueIndex);
+            double destSum = destValue.getDouble(valueIndex + 1);
+            long destCount = destValue.getLong(valueIndex + 2);
+
+            long mergedCount = srcCount + destCount;
+            double delta = destMean - srcMean;
+
+            // This is only valid when countA is much larger than countB.
+            // If both are large and similar sizes, delta is not scaled down.
+            // double mergedMean = srcMean + delta * ((double) destCount / mergedCount);
+
+            // So we use this instead:
+            double mergedMean = (srcCount * srcMean + destCount * destMean) / mergedCount;
+            double mergedSum = srcSum + destSum + (delta * delta) * ((double) (srcCount * destCount) / mergedCount);
+
+            destValue.putDouble(valueIndex, mergedMean);
+            destValue.putDouble(valueIndex + 1, mergedSum);
+            destValue.putLong(valueIndex + 2, mergedCount);
+        }
+
+        @Override
+        public boolean supportsParallelism() {
+            return true;
+        }
+    }
+
 }
